+++ trace_bpfel.go
@@ -8,9 +8,34 @@ import (
 	_ "embed"
 	"fmt"
 	"io"
+	"structs"
 
 	"github.com/cilium/ebpf"
 )
+
+type traceConfig struct {
+	_           structs.HostLayout
+	TxThreshold uint64
+	RxThreshold uint64
+}
+
+type traceConnInfo struct {
+	_        structs.HostLayout
+	RxBytes  uint64
+	TxBytes  uint64
+	WarnedRx uint32
+	WarnedTx uint32
+	Daddr    uint32
+	Dport    uint16
+	Family   uint16
+	Comm     [16]int8
+}
+
+type traceConnKey struct {
+	_   structs.HostLayout
+	Pid uint32
+	Fd  uint32
+}
 
 // loadTrace returns the embedded CollectionSpec for trace.
 func loadTrace() (*ebpf.CollectionSpec, error) {
@@ -54,11 +79,20 @@ type traceSpecs struct {
 //
 // It can be passed ebpf.CollectionSpec.Assign.
 type traceProgramSpecs struct {
-	TracepointSysEnterAccept4 *ebpf.ProgramSpec `ebpf:"tracepoint__sys_enter_accept4"`
-	TracepointSysEnterConnect *ebpf.ProgramSpec `ebpf:"tracepoint__sys_enter_connect"`
-	TracepointSysEnterExecve  *ebpf.ProgramSpec `ebpf:"tracepoint__sys_enter_execve"`
-	TracepointSysExitAccept4  *ebpf.ProgramSpec `ebpf:"tracepoint__sys_exit_accept4"`
-	TracepointSysExitExecve   *ebpf.ProgramSpec `ebpf:"tracepoint__sys_exit_execve"`
+	TracepointSysEnterAccept4  *ebpf.ProgramSpec `ebpf:"tracepoint__sys_enter_accept4"`
+	TracepointSysEnterClose    *ebpf.ProgramSpec `ebpf:"tracepoint__sys_enter_close"`
+	TracepointSysEnterConnect  *ebpf.ProgramSpec `ebpf:"tracepoint__sys_enter_connect"`
+	TracepointSysEnterExecve   *ebpf.ProgramSpec `ebpf:"tracepoint__sys_enter_execve"`
+	TracepointSysEnterRead     *ebpf.ProgramSpec `ebpf:"tracepoint__sys_enter_read"`
+	TracepointSysEnterRecvfrom *ebpf.ProgramSpec `ebpf:"tracepoint__sys_enter_recvfrom"`
+	TracepointSysEnterSendto   *ebpf.ProgramSpec `ebpf:"tracepoint__sys_enter_sendto"`
+	TracepointSysEnterWrite    *ebpf.ProgramSpec `ebpf:"tracepoint__sys_enter_write"`
+	TracepointSysExitAccept4   *ebpf.ProgramSpec `ebpf:"tracepoint__sys_exit_accept4"`
+	TracepointSysExitExecve    *ebpf.ProgramSpec `ebpf:"tracepoint__sys_exit_execve"`
+	TracepointSysExitRead      *ebpf.ProgramSpec `ebpf:"tracepoint__sys_exit_read"`
+	TracepointSysExitRecvfrom  *ebpf.ProgramSpec `ebpf:"tracepoint__sys_exit_recvfrom"`
+	TracepointSysExitSendto    *ebpf.ProgramSpec `ebpf:"tracepoint__sys_exit_sendto"`
+	TracepointSysExitWrite     *ebpf.ProgramSpec `ebpf:"tracepoint__sys_exit_write"`
 }
 
 // traceMapSpecs contains maps before they are loaded into the kernel.
@@ -66,6 +100,9 @@ type traceProgramSpecs struct {
 // It can be passed ebpf.CollectionSpec.Assign.
 type traceMapSpecs struct {
 	Accept4Sockaddr *ebpf.MapSpec `ebpf:"accept4_sockaddr"`
+	ConfigMap       *ebpf.MapSpec `ebpf:"config_map"`
+	ConnStatsMap    *ebpf.MapSpec `ebpf:"conn_stats_map"`
+	IoEnterArgs     *ebpf.MapSpec `ebpf:"io_enter_args"`
 	Rb              *ebpf.MapSpec `ebpf:"rb"`
 }
 
@@ -96,12 +133,18 @@ func (o *traceObjects) Close() error {
 // It can be passed to loadTraceObjects or ebpf.CollectionSpec.LoadAndAssign.
 type traceMaps struct {
 	Accept4Sockaddr *ebpf.Map `ebpf:"accept4_sockaddr"`
+	ConfigMap       *ebpf.Map `ebpf:"config_map"`
+	ConnStatsMap    *ebpf.Map `ebpf:"conn_stats_map"`
+	IoEnterArgs     *ebpf.Map `ebpf:"io_enter_args"`
 	Rb              *ebpf.Map `ebpf:"rb"`
 }
 
 func (m *traceMaps) Close() error {
 	return _TraceClose(
 		m.Accept4Sockaddr,
+		m.ConfigMap,
+		m.ConnStatsMap,
+		m.IoEnterArgs,
 		m.Rb,
 	)
 }
@@ -116,20 +159,38 @@ type traceVariables struct {
 //
 // It can be passed to loadTraceObjects or ebpf.CollectionSpec.LoadAndAssign.
 type tracePrograms struct {
-	TracepointSysEnterAccept4 *ebpf.Program `ebpf:"tracepoint__sys_enter_accept4"`
-	TracepointSysEnterConnect *ebpf.Program `ebpf:"tracepoint__sys_enter_connect"`
-	TracepointSysEnterExecve  *ebpf.Program `ebpf:"tracepoint__sys_enter_execve"`
-	TracepointSysExitAccept4  *ebpf.Program `ebpf:"tracepoint__sys_exit_accept4"`
-	TracepointSysExitExecve   *ebpf.Program `ebpf:"tracepoint__sys_exit_execve"`
+	TracepointSysEnterAccept4  *ebpf.Program `ebpf:"tracepoint__sys_enter_accept4"`
+	TracepointSysEnterClose    *ebpf.Program `ebpf:"tracepoint__sys_enter_close"`
+	TracepointSysEnterConnect  *ebpf.Program `ebpf:"tracepoint__sys_enter_connect"`
+	TracepointSysEnterExecve   *ebpf.Program `ebpf:"tracepoint__sys_enter_execve"`
+	TracepointSysEnterRead     *ebpf.Program `ebpf:"tracepoint__sys_enter_read"`
+	TracepointSysEnterRecvfrom *ebpf.Program `ebpf:"tracepoint__sys_enter_recvfrom"`
+	TracepointSysEnterSendto   *ebpf.Program `ebpf:"tracepoint__sys_enter_sendto"`
+	TracepointSysEnterWrite    *ebpf.Program `ebpf:"tracepoint__sys_enter_write"`
+	TracepointSysExitAccept4   *ebpf.Program `ebpf:"tracepoint__sys_exit_accept4"`
+	TracepointSysExitExecve    *ebpf.Program `ebpf:"tracepoint__sys_exit_execve"`
+	TracepointSysExitRead      *ebpf.Program `ebpf:"tracepoint__sys_exit_read"`
+	TracepointSysExitRecvfrom  *ebpf.Program `ebpf:"tracepoint__sys_exit_recvfrom"`
+	TracepointSysExitSendto    *ebpf.Program `ebpf:"tracepoint__sys_exit_sendto"`
+	TracepointSysExitWrite     *ebpf.Program `ebpf:"tracepoint__sys_exit_write"`
 }
 
 func (p *tracePrograms) Close() error {
 	return _TraceClose(
 		p.TracepointSysEnterAccept4,
+		p.TracepointSysEnterClose,
 		p.TracepointSysEnterConnect,
 		p.TracepointSysEnterExecve,
+		p.TracepointSysEnterRead,
+		p.TracepointSysEnterRecvfrom,
+		p.TracepointSysEnterSendto,
+		p.TracepointSysEnterWrite,
 		p.TracepointSysExitAccept4,
 		p.TracepointSysExitExecve,
+		p.TracepointSysExitRead,
+		p.TracepointSysExitRecvfrom,
+		p.TracepointSysExitSendto,
+		p.TracepointSysExitWrite,
 	)
 }